Max length
Found out the issue with the Easy selection is a limitation in the `RandomWords` module. Max length needs to be at least 5 or greater characters. Resolved this issue by bumping the character limit from 4 to 5. Also resolved an issue where if someone types in the whole word, the app does not know what to do. Now, we take the first character instead.